Fix French string translations not showing up on website
Ensure the WPML MO files are refreshed with the latest string translations
in the WordPress database when we export the website with Simply Static
by adding:

    define( 'WPML_ST_SYNC_TRANSLATION_FILES', true );

to wp-config-docker.php (and hence the autogenerated wp-config.php.

Not intended for live site due to performance penalty.

See https://wpml.org/forums/topic/rebuild-all-mo-files-from-the-string-translation-database/
and https://wpml.org/documentation/getting-started-guide/string-translation/finding-strings-that-dont-appear-on-the-string-translation-page/#synchronizing-translation-files-for-sites-running-on-multiple-servers

Fixes OpenDRR/riskprofiler#114